requirements.txt
flask~=2.2.2
python-dotenv~=0.21.0
openai~=0.27.4
waitress~=2.1.1